requirements.txt
argparse
easydict
h5py
matplotlib
numpy
open3d
opencv-python
pyexr
scipy
tensorboardX==1.2
torch==1.4.0
transforms3d
tqdm
torchvision==0.5.0
yacs
scikit-image
Pillow
pyyaml
efficientnet_pytorch
colorlog
tensorboard
openpyxl